remove what should be subtasks
also align naming more with document and update some of
the arrows indicating we may update data in a subtask

Additionally, the spec states repetitively that the search screen should
be the same, but have different functionality displayed depending on the
user. This we should cover in the abstract code.