add more test data and fixup sellers history report
The Sellers History report that I worked on had a couple issues.
We were missing vehicles with no parts orders, and I incorporated the
highlighting logic so we could just key off that in front end.

Also move logic to update total_parts_price to the "load_data.sql" as we
will have to do this in the application, not with triggers etc per
response on piazza.